SOC-650 | refactor: extract parent class from forgot-password and reset-password classes
