+++ front/auth/app/reset-password/reset-password.js
@@ -1,118 +1,78 @@
-import AuthTracker from '../common/auth-tracker';
-import AuthLogger from '../common/auth-logger';
 import HttpCodes from '../common/http-codes';
-import UrlHelper from '../common/url-helper';
+import PasswordForm from '../common/password-form';
 import {track as mercuryTrack, trackActions} from 'common/utils/track';
 
-/**
- * @typedef {Object} FormElements
- * @property {HTMLInputElement} username
- */
+export default class ResetPassword extends PasswordForm {
 
-/**
- * @class Login
- *
- * @property {HTMLFormElement} form
- * @property {string} redirect
- * @property {HTMLInputElement} usernameInput
- * @property {UrlHelper} urlHelper
- * @property {AuthTracker} tracker
- * @property {AuthLogger} authLogger
- */
-export default class ResetPassword {
 	/**
 	 * @param {Element} form
-	 * @returns {void}
 	 */
 	constructor(form) {
-		this.form = form;
-		this.newPasswordInput = form.elements.newPassword;
-		this.confirmNewPasswordInput = form.elements.confirmNewPassword;
-
-		this.authLogger = AuthLogger.getInstance();
-		this.tracker = new AuthTracker('reset-password-mobile', '/resetPassword');
-		this.urlHelper = new UrlHelper();
-
-		this.extractParametersFromUrlQuery();
+		super(form, {
+			category: 'reset-password-mobile',
+			pageType: '/resetPassword'
+		});
 	}
 
 	/**
-	 * @private
-	 */
-	extractParametersFromUrlQuery() {
-		if (window.location.search) {
-			const params = this.urlHelper.urlDecode(window.location.search.substr(1));
-			this.redirect = params.redirect;
-			this.token = params.token;
-			this.username = params.username;
-		}
-	}
-
-	/**
-	 * @param {Event} event
+	 * @protected
 	 *
-	 * @returns {void}
+	 * @param {array} parameters
 	 */
-	onSubmit(event) {
-		event.preventDefault();
+	extractFieldsFromPathParameters(parameters) {
+		this.redirect = parameters.redirect;
+		this.token = parameters.token;
+		this.username = parameters.username;
+	}
 
-		const button = this.form.querySelector('button'),
-			newPasswordValue = this.newPasswordInput.value,
-			confirmNewPasswordValue = this.confirmNewPasswordInput.value,
-			data = {
-				password: newPasswordValue,
-				redirect: this.redirect,
-				username: this.username
-			},
-			xhr = new XMLHttpRequest();
 
-		this.clearError();
-		button.disabled = true;
+	/**
+	 * @protected
+	 *
+	 * @returns {object} data to be send
+	 */
+	collectDataBeforeSubmit() {
+		return {
+			password: this.form.elements.newPassword.value,
+			redirect: this.redirect,
+			username: this.username
+		};
+	}
 
-		if (newPasswordValue !== confirmNewPasswordValue) {
+	/**
+	 * @protected
+	 *
+	 * @returns {boolean} true if input is validated, false if errors were found
+	 */
+	inputIsValid() {
+		const newPasswordValue = this.form.elements.newPassword.value,
+			confirmNewPasswordValue = this.form.elements.confirmNewPassword.value,
+			passwordIsSame = newPasswordValue === confirmNewPasswordValue;
+
+		if (!passwordIsSame) {
 			this.displayError('errors.passwords_not_match');
-		} else {
-			xhr.onload = () => {
-				button.disabled = false;
-
-				if (xhr.status === HttpCodes.OK) {
-					this.onSuccess();
-				} else {
-					this.handleErrors(xhr);
-				}
-			};
-
-			xhr.onerror = () => {
-				button.disabled = false;
-				this.onError(xhr);
-			};
-
-			xhr.open('post', this.form.action, true);
-			xhr.setRequestHeader('Content-type', 'application/x-www-form-urlencoded');
-			xhr.send(this.urlHelper.urlEncode(data));
 		}
+
+		return passwordIsSame;
 	}
 
-	handleErrors(xhr) {
-		const response = JSON.parse(xhr.responseText);
 
-		if (response.step === 'service-discovery') {
-			this.onError(xhr);
-		} else if (response.step === 'user-discovery') {
-			this.handleUserDiscoveryErrors(xhr);
-		} else if (response.step === 'update-password') {
+	/**
+	 * @protected
+	 *
+	 * @param {XMLHttpRequest} xhr
+	 * @param {object} response - parsed json response
+	 *
+	 * @returns {boolean} - true if errors were handled, false otherwise
+	 */
+	handleCustomErrors(xhr, response) {
+		let errorsHandled = response.step === 'update-password';
+
+		if (errorsHandled) {
 			this.handleUpdatePasswordErrors(xhr, response);
-		} else {
-			this.onError(xhr);
 		}
-	}
 
-	handleUserDiscoveryErrors(xhr) {
-		if (xhr.status === HttpCodes.NOT_FOUND) {
-			this.onUsernameNotRecognizedError();
-		} else {
-			this.onError(xhr);
-		}
+		return errorsHandled;
 	}
 
 	handleUpdatePasswordErrors(xhr, response) {
@@ -134,57 +94,6 @@ export default class ResetPassword {
 
 	hasError(response) {
 		return response && response.errors && response.errors.length;
-	}
-
-	onError(xhr) {
-		this.authLogger.xhrError(xhr);
-		this.tracker.track('server-error', trackActions.error);
-		this.displayError('errors.server-error');
-	}
-
-	onUsernameNotRecognizedError() {
-		this.tracker.track('username-not-recognized', trackActions.error);
-		this.displayError('errors.username-not-recognized');
-	}
-
-	/**
-	 * @returns {void}
-	 */
-	onSuccess() {
-		this.tracker.track('forgot-password-success', trackActions.submit);
-		document.querySelector('.cards-container').classList.add('dissolved');
-	}
-
-	/**
-	 * @returns {void}
-	 */
-	watch() {
-		this.tracker.trackCloseWindow();
-		this.form.addEventListener('submit', this.onSubmit.bind(this));
-	}
-
-	/**
-	 * @param {string} messageKey
-	 *
-	 * @returns {void}
-	 */
-	displayError(messageKey) {
-		const errorElement = document.createElement('small');
-
-		errorElement.classList.add('error');
-		errorElement.innerHTML = i18n.t(messageKey);
-		this.newPasswordInput.parentElement.appendChild(errorElement);
-	}
-
-	/**
-	 * @returns {void}
-	 */
-	clearError() {
-		const errorNode = this.form.querySelector('small.error');
-
-		if (errorNode) {
-			errorNode.parentNode.removeChild(errorNode);
-		}
 	}
 }
